[Wallet]: Merge CryptoView into Unlocked Wallet Routes
