--- components/brave_wallet_ui/components/desktop/views/crypto/index.tsx
@@ -1,226 +1,0 @@
-// Copyright (c) 2022 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
-
-import * as React from 'react'
-import { Route, useHistory, Switch, Redirect } from 'react-router-dom'
-import { useSelector } from 'react-redux'
-
-// types
-import { WalletRoutes } from '../../../../constants/types'
-import {
-  AccountsTabState, //
-} from '../../../../page/reducers/accounts-tab-reducer'
-
-// hooks
-import {
-  usePortfolioVisibleNetworks, //
-} from '../../../../common/hooks/use_portfolio_networks'
-import {
-  usePortfolioAccounts, //
-} from '../../../../common/hooks/use_portfolio_accounts'
-
-// style
-import { DefaultPageWrapper } from '../../../shared/style'
-
-// components
-import { ExploreWeb3Header } from '../../card-headers/explorer_web3_header'
-import {
-  EditVisibleAssetsModal, //
-} from '../../popup-modals/edit-visible-assets-modal/index'
-import { PortfolioOverview } from '../../../../page/screens/portfolio_overview/portfolio_overview'
-import { FungibleAssetDetails } from '../../../../page/screens/fungible_asset_details/fungible_asset_details'
-import { NFTAssetDetails } from '../../../../page/screens/nft_asset_details/nft_asset_details'
-import { MarketView } from '../../../../page/screens/market/market'
-import { AccountsOverview } from '../../../../page/screens/accounts_overview/accounts_overview'
-import { AccountDetails } from '../../../../page/screens/account_details/account_details'
-import { AddAccountModal } from '../../popup-modals/add-account-modal/add-account-modal'
-import {
-  RemoveAccountModal, //
-} from '../../popup-modals/confirm-password-modal/remove-account-modal'
-import { AccountSettingsModal } from '../../popup-modals/account-settings-modal/account-settings-modal'
-import {
-  WalletPageWrapper, //
-} from '../../wallet-page-wrapper/wallet-page-wrapper'
-import { MarketAssetDetails } from '../../../../page/screens/market/market_asset_details'
-import { NftCollection } from '../nfts/components/nft_collection'
-import {
-  PageNotFound, //
-} from '../../../../page/screens/page_not_found/page_not_found'
-import { Banners } from '../banners/banners'
-
-export interface Props {
-  sessionRoute: string | undefined
-}
-
-export const CryptoView = ({ sessionRoute }: Props) => {
-  // Selectors
-  const { accountToRemove, showAccountModal, selectedAccount } = useSelector(
-    ({ accountsTab }: { accountsTab: AccountsTabState }) => accountsTab,
-  )
-
-  // custom hooks
-  const { visiblePortfolioNetworks } = usePortfolioVisibleNetworks()
-  const { usersFilteredAccounts } = usePortfolioAccounts()
-
-  // routing
-  const history = useHistory()
-
-  // methods
-  const onShowVisibleAssetsModal = React.useCallback(
-    (showModal: boolean) => {
-      if (showModal) {
-        history.push(WalletRoutes.AddAssetModal)
-        return
-      }
-      history.push(WalletRoutes.PortfolioAssets)
-    },
-    [history],
-  )
-
-  const hideVisibleAssetsModal = React.useCallback(
-    () => onShowVisibleAssetsModal(false),
-    [onShowVisibleAssetsModal],
-  )
-
-  // render
-  return (
-    <>
-      <Switch>
-        {/* Portfolio */}
-        <Route
-          path={WalletRoutes.Portfolio}
-          exact={true}
-          render={() => <Redirect to={WalletRoutes.PortfolioAssets} />}
-        />
-
-        <Route
-          path={WalletRoutes.PortfolioAssets}
-          exact
-        >
-          <PortfolioOverview />
-        </Route>
-
-        <Route
-          path={WalletRoutes.PortfolioNFTs}
-          exact
-        >
-          <PortfolioOverview />
-        </Route>
-
-        <Route
-          path={WalletRoutes.PortfolioActivity}
-          exact
-        >
-          <PortfolioOverview />
-        </Route>
-
-        <Route path={WalletRoutes.AddAssetModal}>
-          <PortfolioOverview />
-        </Route>
-
-        <Route
-          path={WalletRoutes.PortfolioNFTAsset}
-          exact
-        >
-          <NFTAssetDetails />
-        </Route>
-
-        <Route
-          path={WalletRoutes.PortfolioAsset}
-          exact
-        >
-          <FungibleAssetDetails />
-        </Route>
-
-        <Route
-          path={WalletRoutes.PortfolioNFTCollection}
-          exact
-        >
-          <NftCollection
-            networks={visiblePortfolioNetworks}
-            accounts={usersFilteredAccounts}
-          />
-        </Route>
-
-        {/* Accounts */}
-        <Route path={WalletRoutes.AddAccountModal}>
-          {/* Show accounts overview in background */}
-          <WalletPageWrapper wrapContentInBox={true}>
-            <DefaultPageWrapper>
-              <Banners />
-              <AccountsOverview />
-            </DefaultPageWrapper>
-          </WalletPageWrapper>
-        </Route>
-
-        <Route path={WalletRoutes.Account}>
-          <AccountDetails />
-        </Route>
-
-        <Route path={WalletRoutes.Accounts}>
-          <AccountsOverview />
-        </Route>
-
-        {/* Market */}
-        <Route
-          path={WalletRoutes.Market}
-          exact={true}
-        >
-          <WalletPageWrapper
-            wrapContentInBox
-            cardHeader={<ExploreWeb3Header />}
-            useCardInPanel={true}
-          >
-            <DefaultPageWrapper>
-              <MarketView />
-            </DefaultPageWrapper>
-          </WalletPageWrapper>
-        </Route>
-
-        <Route
-          path={WalletRoutes.MarketSub}
-          exact={true}
-        >
-          <WalletPageWrapper wrapContentInBox={true}>
-            <DefaultPageWrapper>
-              <MarketAssetDetails />
-            </DefaultPageWrapper>
-          </WalletPageWrapper>
-        </Route>
-
-        <Route
-          path={WalletRoutes.Explore}
-          exact={true}
-        >
-          <Redirect to={WalletRoutes.Market} />
-        </Route>
-        <Route path='*'>
-          <PageNotFound />
-        </Route>
-        <Redirect to={sessionRoute ?? WalletRoutes.PortfolioAssets} />
-      </Switch>
-
-      {/* modals */}
-      <Switch>
-        <Route
-          path={WalletRoutes.AddAssetModal}
-          exact
-        >
-          <EditVisibleAssetsModal onClose={hideVisibleAssetsModal} />
-        </Route>
-
-        <Route path={WalletRoutes.AddAccountModal}>
-          <AddAccountModal />
-        </Route>
-      </Switch>
-
-      {accountToRemove !== undefined && <RemoveAccountModal />}
-
-      {showAccountModal && selectedAccount && <AccountSettingsModal />}
-    </>
-  )
-}
-
-export default CryptoView
